reorg projects and dependencies a bit add tsconfig-paths to ts-node execution for api
