+++ api/package.json
@@ -20,6 +20,7 @@
     "pg": "^8.11.1",
     "pg-native": "^3.0.1",
     "reflect-metadata": "^0.1.13",
+    "ts-node": "^10.9.1",
     "typeorm": "^0.3.17",
     "typescript": "^5.1.6"
   },
@@ -28,6 +29,7 @@
     "@typescript-eslint/parser": "^5.60.1",
     "eslint": "^8.43.0",
     "eslint-config-prettier": "^8.8.0",
-    "prettier": "^2.8.8"
+    "prettier": "^2.8.8",
+    "tsconfig-paths": "^4.2.0"
   }
 }
