+++ next/tsconfig.json
@@ -26,16 +26,6 @@
     "strictPropertyInitialization": false,
     "experimentalDecorators": true
   },
-  "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx",
-    ".next/types/**/*.ts",
-    "../models/src/",
-    "../api/src/entity",
-    "../api/src/migration",
-    "../api/src/route.ts",
-    "../api/src/API.ts"
-  ],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
   "exclude": ["node_modules"]
 }
